requirements.txt
streamlit~=1.29.0
pydantic~=2.5.2
Pypubsub~=4.0.3
SQLAlchemy~=2.0.23
pandas~=2.1.4
openai~=1.3.8
yfinance~=0.2.33
transitions~=0.9.0
python-dotenv~=1.0.0